+++ licensecheck/checker.py
@@ -3,50 +3,18 @@
 from __future__ import annotations
 
 from fnmatch import fnmatch
-from pathlib import Path
 
-import tomli
-from loguru import logger
 
 from licensecheck import license_matrix
 from licensecheck.packageinfo import PackageInfoManager
-from licensecheck.resolvers import native as res_native
-from licensecheck.resolvers import uv as res_uv
 from licensecheck.types import JOINS, License, PackageInfo, ucstr
 
-
-def resolve_requirements(
-	requirements_paths: list[str],
-	groups: list[str],
-	skip_dependencies: list[ucstr],
-	index_url: str,
-) -> set[ucstr]:
-	try:
-		return res_uv.get_reqs(
-			skipDependencies=skip_dependencies,
-			extras=groups,
-			requirementsPaths=requirements_paths,
-			index_url=index_url,
-		)
-
-	except RuntimeError as e:
-		logger.warning(e)
-		pyproject = {}
-		if "pyproject.toml" in requirements_paths:
-			pyproject = tomli.loads(Path("pyproject.toml").read_text("utf-8"))
-
-		# Fallback to the old resolver (hopefully we can deprecate this asap!)
-		return res_native.get_reqs(
-			skipDependencies=skip_dependencies,
-			extras=groups,
-			pyproject=pyproject,
-			requirementsPaths=[Path(x) for x in requirements_paths],
-		)
 
 
 def check(
 	requirements_paths: list[str],
 	groups: list[str],
+	extras: list[str],
 	this_license: License,
 	package_info_manager: PackageInfoManager,
 	ignore_packages: list[ucstr] | None = None,
@@ -64,11 +32,11 @@ def check(
 	only_licenses = only_licenses or []
 	skip_dependencies = skip_dependencies or []
 
-	requirements = resolve_requirements(
-		requirements_paths,
-		groups,
-		skip_dependencies,
-		index_url=package_info_manager.pypi_search,
+	package_info_manager.resolve_requirements(
+		requirements_paths=requirements_paths,
+		groups=groups,
+		extras=extras,
+		skip_dependencies=skip_dependencies,
 	)
 
 	ignoreLicensesType = license_matrix.licenseType(
@@ -81,7 +49,7 @@ def check(
 		onlyLicensesType.remove(License.NO_LICENSE)
 
 	# Check it is compatible with packages and add a note
-	packages = package_info_manager.getPackages(requirements)
+	packages = package_info_manager.getPackages()
 	for package in packages:
 		# Deal with --ignore-packages and --fail-packages
 		package.licenseCompat = False
@@ -94,7 +62,7 @@ def check(
 		else:
 			package.licenseCompat = license_matrix.depCompatWMyLice(
 				this_license,
-				license_matrix.licenseType(package.license, ignore_licenses),
+				license_matrix.licenseType(ucstr(package.license), ignore_licenses),
 				ignoreLicensesType,
 				failLicensesType,
 				onlyLicensesType,