+++ tests/platform_independent/test_packageinfo.py
@@ -6,54 +6,69 @@ from typing import Any
 
 import pytest
 
-from licensecheck import packageinfo, types
-from licensecheck.packageinfo import PackageInfoManager
+from licensecheck.packageinfo import LocalPackageInfo, RemotePackageInfo,PackageInfoManager, from_classifiers, meta_get
+from licensecheck.types import PackageInfo, UNKNOWN
 
 THISDIR = str(Path(__file__).resolve().parent)
 
 
-package_info_manager = PackageInfoManager("https://pypi.org/pypi/")
+@pytest.fixture
+def package_info_manager():
+    """Fixture to provide a PackageInfoManager instance."""
+    return PackageInfoManager("https://pypi.org/pypi/")
 
 
-def test_getPackageInfoLocal() -> None:
+@pytest.fixture
+def local_package_info():
+	return LocalPackageInfo(requests_package)
+
+
+@pytest.fixture
+def remote_package_info():
+	return RemotePackageInfo("https://pypi.org/pypi/", requests_package)
+
+
+
+def aux_packageinfo(package_name: str)-> PackageInfo:
+	return PackageInfo(name=package_name)
+
+requests_package = aux_packageinfo("requests")
+
+
+def test_getPackageInfoLocal(local_package_info: LocalPackageInfo) -> None:
 	try:
-		package = packageinfo.LocalPackageInfo.get_info(types.ucstr("requests"))
-		assert package.name == "requests"
+		name = local_package_info.get_name()
+		assert name == "requests"
 	except ModuleNotFoundError:
 		assert True
 
 
-def test_getPackageInfoPypi() -> None:
-	package = packageinfo.RemotePackageInfo.get_info(
-		types.ucstr("requests"), "https://pypi.org/pypi/"
-	)
+def test_getPackageInfoPypi(remote_package_info: RemotePackageInfo) -> None:
 
-	assert package.name == "requests"
-	assert package.homePage == "https://requests.readthedocs.io"
-	assert package.author == "Kenneth Reitz"
-	assert package.license == "APACHE SOFTWARE LICENSE"
+	pkg = remote_package_info
+
+	assert pkg.get_name() == "requests"
+	assert pkg.get_homePage() == "https://requests.readthedocs.io"
+	assert pkg.get_author() == "Kenneth Reitz"
+	assert pkg.get_license() == "Apache Software License"
 
 
 def test_getPackageInfoLocalNotFound() -> None:
-	try:
-		packageinfo.LocalPackageInfo.get_info(types.ucstr("this_package_does_not_exist"))
-		raise AssertionError
-	except ModuleNotFoundError:
-		assert True
+	pkg = LocalPackageInfo(aux_packageinfo("this_package_does_not_exist"))
+	assert pkg.get_size() == 0
 
 
 def test_getPackagePypiLocalNotFound() -> None:
-	try:
-		packageinfo.RemotePackageInfo.get_info(
-			types.ucstr("this_package_does_not_exist"), "https://pypi.org/pypi/"
+	pkg = RemotePackageInfo("https://pypi.org/pypi/",
+			aux_packageinfo("this_package_does_not_exist")
 		)
-		raise AssertionError
-	except ModuleNotFoundError:
-		assert True
+	assert pkg.get_size() == -1
 
 
-def test_getPackages() -> None:
-	packages = package_info_manager.getPackages({types.ucstr("requests")})
+
+def test_getPackages(package_info_manager: PackageInfoManager) -> None:
+	package_info_manager.reqs = {aux_packageinfo("requests")}
+	packages = package_info_manager.getPackages()
 	assert all(
 		(
 			package.name == "requests"
@@ -65,11 +80,13 @@ def test_getPackages() -> None:
 	)
 
 
-def test_getPackagesNotFound() -> None:
-	packages = package_info_manager.getPackages({types.ucstr("this_package_does_not_exist")})
+def test_getPackagesNotFound(package_info_manager: PackageInfoManager) -> None:
+	package_info_manager.reqs = {aux_packageinfo("this_package_does_not_exist")}
+
+	packages = package_info_manager.getPackages()
 
 	assert all(
-		(package.name == "THIS_PACKAGE_DOES_NOT_EXIST" and package.errorCode == 1)
+		(package.name.upper() == "THIS_PACKAGE_DOES_NOT_EXIST" and package.errorCode == 1)
 		for package in packages
 	)
 
@@ -77,20 +94,21 @@ def test_getPackagesNotFound() -> None:
 def test_from_classifiers() -> None:
 	licenses = []
 	for rawLicense in Path(f"{THISDIR}/data/pypiClassifiers.txt").read_text("utf-8").splitlines():
-		licenses.append(packageinfo.from_classifiers([rawLicense]))
+		licenses.append(from_classifiers([rawLicense]) or UNKNOWN)
 	# Path(f"{THISDIR}/data/licenses.txt").write_text("\n".join(licenses), "utf-8")
 	assert "\n".join(licenses) == Path(f"{THISDIR}/data/licenses.txt").read_text("utf-8")
 
 
 def test_licenseFromEmptyClassifierlist() -> None:
 	licenses = []
-	licenses.append(packageinfo.from_classifiers([]))
-	assert licenses == [types.UNKNOWN]
+	licenses.append(from_classifiers([]))
+	assert licenses == [None]
 
 
 def test_getModuleSize() -> None:
-	with pytest.raises(importlib.metadata.PackageNotFoundError):
-		packageinfo.LocalPackageInfo.get_size(types.ucstr("this_package_does_not_exist"))
+	local_package_info = LocalPackageInfo(aux_packageinfo("this_package_does_not_exist"))
+	local_package_info.get_size()
+
 
 
 # Define test cases
@@ -101,11 +119,11 @@ def test_getModuleSize() -> None:
 		({"name": ["Package Name"], "version": "1.0"}, "name", "Package Name"),
 		({"name": [1], "version": "1.0"}, "name", "1"),
 		({"name": 1, "version": "1.0"}, "name", "1"),
-		({"name": None, "version": "1.0"}, "name", types.UNKNOWN),
+		({"name": None, "version": "1.0"}, "name", None),
 		({"name": ["Package", "Name"], "version": "1.0"}, "name", "Package AND Name"),
-		({}, "name", types.UNKNOWN),
-		({"name": "Package Name", "version": "1.0"}, "description", types.UNKNOWN),
+		({}, "name", None),
+		({"name": "Package Name", "version": "1.0"}, "description", None),
 	],
 )
 def test_get_metadata(pkg_metadata: dict[str, Any], key: str, expected: str) -> None:
-	assert packageinfo.meta_get(pkg_metadata, key) == expected
+	assert meta_get(pkg_metadata, key) == expected