+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2025.1.0 - 2025/03/18
+
+- Updated internal machineary
+- Separate `--extras` and `--groups`
+
 ## 2025.0.1 - 2025/03/16
 
 - update the cli interface to remove some confusing functionality